Auto-Subtitled-Video-Generator/requirements.txt
ffmpeg-python==0.2.0
numpy>=1.24.0
pytubefix==6.13.0
requests==2.32.3
streamlit==1.37.1
streamlit_lottie==0.0.3
torch>=2.0.0
transformers>=4.30.0
openai-whisper